CloudMonitor: remove cloudMonitoringExperimentalUI
