+++ packages/grafana-data/src/types/featureToggles.gen.ts
@@ -57,7 +57,6 @@ export interface FeatureToggles {
   canvasPanelNesting?: boolean;
   scenes?: boolean;
   useLegacyHeatmapPanel?: boolean;
-  cloudMonitoringExperimentalUI?: boolean;
   disableSecretsCompatibility?: boolean;
   logRequestsInstrumentedAsUnknown?: boolean;
   dataConnectionsConsole?: boolean;
