+++ public/app/plugins/datasource/cloud-monitoring/components/Aggregation.tsx
@@ -1,12 +1,10 @@
 import React, { FC, useMemo } from 'react';
 
 import { SelectableValue } from '@grafana/data';
-import { Select } from '@grafana/ui';
+import { EditorField, Select } from '@grafana/ui';
 
 import { getAggregationOptionsByMetric } from '../functions';
 import { MetricDescriptor, MetricKind, ValueTypes } from '../types';
-
-import { QueryEditorField } from '.';
 
 export interface Props {
   refId: string;
@@ -22,14 +20,9 @@ export const Aggregation: FC<Props> = (props) => {
   const selected = useSelectedFromOptions(aggOptions, props);
 
   return (
-    <QueryEditorField
-      labelWidth={18}
-      label="Group by function"
-      data-testid="cloud-monitoring-aggregation"
-      htmlFor={`${props.refId}-group-by-function`}
-    >
+    <EditorField label="Group by function" data-testid="cloud-monitoring-aggregation">
       <Select
-        width={16}
+        width="auto"
         onChange={({ value }) => props.onChange(value!)}
         value={selected}
         options={[
@@ -46,7 +39,7 @@ export const Aggregation: FC<Props> = (props) => {
         placeholder="Select Reducer"
         inputId={`${props.refId}-group-by-function`}
       />
-    </QueryEditorField>
+    </EditorField>
   );
 };
