+++ public/app/plugins/datasource/cloud-monitoring/components/AliasBy.tsx
@@ -1,11 +1,9 @@
 import { debounce } from 'lodash';
 import React, { FunctionComponent, useState } from 'react';
 
-import { Input } from '@grafana/ui';
+import { EditorField, Input } from '@grafana/ui';
 
-import { INPUT_WIDTH } from '../constants';
-
-import { QueryEditorRow } from '.';
+import { SELECT_WIDTH } from '../constants';
 
 export interface Props {
   refId: string;
@@ -24,8 +22,8 @@ export const AliasBy: FunctionComponent<Props> = ({ refId, value = '', onChange 
   };
 
   return (
-    <QueryEditorRow label="Alias by" htmlFor={`${refId}-alias-by`}>
-      <Input id={`${refId}-alias-by`} width={INPUT_WIDTH} value={alias} onChange={onChange} />
-    </QueryEditorRow>
+    <EditorField label="Alias by">
+      <Input id={`${refId}-alias-by`} width={SELECT_WIDTH} value={alias} onChange={onChange} />
+    </EditorField>
   );
 };
