+++ public/app/plugins/datasource/cloud-monitoring/components/Alignment.tsx
@@ -1,12 +1,15 @@
-import React, { FC } from 'react';
+import React, { FC, useMemo } from 'react';
 
 import { SelectableValue } from '@grafana/data';
+import { EditorField, EditorFieldGroup } from '@grafana/ui';
 
-import { ALIGNMENT_PERIODS, SELECT_WIDTH } from '../constants';
+import { ALIGNMENT_PERIODS } from '../constants';
 import CloudMonitoringDatasource from '../datasource';
+import { alignmentPeriodLabel } from '../functions';
 import { CustomMetaData, MetricQuery, SLOQuery } from '../types';
 
-import { AlignmentFunction, PeriodSelect, AlignmentPeriodLabel, QueryEditorField, QueryEditorRow } from '.';
+import { AlignmentFunction } from './AlignmentFunction';
+import { PeriodSelect } from './PeriodSelect';
 
 export interface Props {
   refId: string;
@@ -25,29 +28,29 @@ export const Alignment: FC<Props> = ({
   customMetaData,
   datasource,
 }) => {
+  const alignmentLabel = useMemo(() => alignmentPeriodLabel(customMetaData, datasource), [customMetaData, datasource]);
   return (
-    <QueryEditorRow
-      label="Alignment function"
-      tooltip="The process of alignment consists of collecting all data points received in a fixed length of time, applying a function to combine those data points, and assigning a timestamp to the result."
-      fillComponent={<AlignmentPeriodLabel datasource={datasource} customMetaData={customMetaData} />}
-      htmlFor={`${refId}-alignment-function`}
-    >
-      <AlignmentFunction
-        inputId={`${refId}-alignment-function`}
-        templateVariableOptions={templateVariableOptions}
-        query={query}
-        onChange={onChange}
-      />
-      <QueryEditorField label="Alignment period" htmlFor={`${refId}-alignment-period`}>
+    <EditorFieldGroup>
+      <EditorField
+        label="Alignment function"
+        tooltip="The process of alignment consists of collecting all data points received in a fixed length of time, applying a function to combine those data points, and assigning a timestamp to the result."
+      >
+        <AlignmentFunction
+          inputId={`${refId}-alignment-function`}
+          templateVariableOptions={templateVariableOptions}
+          query={query}
+          onChange={onChange}
+        />
+      </EditorField>
+      <EditorField label="Alignment period" tooltip={alignmentLabel}>
         <PeriodSelect
           inputId={`${refId}-alignment-period`}
-          selectWidth={SELECT_WIDTH}
           templateVariableOptions={templateVariableOptions}
           current={query.alignmentPeriod}
           onChange={(period) => onChange({ ...query, alignmentPeriod: period })}
           aligmentPeriods={ALIGNMENT_PERIODS}
         />
-      </QueryEditorField>
-    </QueryEditorRow>
+      </EditorField>
+    </EditorFieldGroup>
   );
 };
