+++ public/app/plugins/datasource/cloud-monitoring/components/AnnotationQueryEditor.tsx
@@ -2,10 +2,8 @@ import React, { useState } from 'react';
 import { useDebounce } from 'react-use';
 
 import { QueryEditorProps, toOption } from '@grafana/data';
-import { config } from '@grafana/runtime';
 import { EditorField, EditorRows, Input } from '@grafana/ui';
 
-import { INPUT_WIDTH } from '../constants';
 import CloudMonitoringDatasource from '../datasource';
 import {
   EditorMode,
@@ -16,10 +14,9 @@ import {
   AlignmentTypes,
 } from '../types';
 
-import { MetricQueryEditor as ExperimentalMetricQueryEditor } from './Experimental/MetricQueryEditor';
 import { MetricQueryEditor } from './MetricQueryEditor';
 
-import { AnnotationsHelp, QueryEditorRow } from './';
+import { AnnotationsHelp } from './';
 
 export type Props = QueryEditorProps<CloudMonitoringDatasource, CloudMonitoringQuery, CloudMonitoringOptions>;
 
@@ -80,44 +77,23 @@ export const AnnotationQueryEditor = (props: Props) => {
 
   return (
     <EditorRows>
-      {config.featureToggles.cloudMonitoringExperimentalUI ? (
-        <>
-          <ExperimentalMetricQueryEditor
-            refId={query.refId}
-            variableOptionGroup={variableOptionGroup}
-            customMetaData={customMetaData}
-            onChange={handleQueryChange}
-            onRunQuery={onRunQuery}
-            datasource={datasource}
-            query={metricQuery}
-          />
-          <EditorField label="Title" htmlFor="annotation-query-title">
-            <Input id="annotation-query-title" value={title} onChange={handleTitleChange} />
-          </EditorField>
-          <EditorField label="Text" htmlFor="annotation-query-text">
-            <Input id="annotation-query-text" value={text} onChange={handleTextChange} />
-          </EditorField>
-        </>
-      ) : (
-        <>
-          <MetricQueryEditor
-            refId={query.refId}
-            variableOptionGroup={variableOptionGroup}
-            customMetaData={customMetaData}
-            onChange={handleQueryChange}
-            onRunQuery={onRunQuery}
-            datasource={datasource}
-            query={metricQuery}
-          />
-          <QueryEditorRow label="Title" htmlFor="annotation-query-title">
-            <Input id="annotation-query-title" value={title} width={INPUT_WIDTH} onChange={handleTitleChange} />
-          </QueryEditorRow>
-
-          <QueryEditorRow label="Text" htmlFor="annotation-query-text">
-            <Input id="annotation-query-text" value={text} width={INPUT_WIDTH} onChange={handleTextChange} />
-          </QueryEditorRow>
-        </>
-      )}
+      <>
+        <MetricQueryEditor
+          refId={query.refId}
+          variableOptionGroup={variableOptionGroup}
+          customMetaData={customMetaData}
+          onChange={handleQueryChange}
+          onRunQuery={onRunQuery}
+          datasource={datasource}
+          query={metricQuery}
+        />
+        <EditorField label="Title" htmlFor="annotation-query-title">
+          <Input id="annotation-query-title" value={title} onChange={handleTitleChange} />
+        </EditorField>
+        <EditorField label="Text" htmlFor="annotation-query-text">
+          <Input id="annotation-query-text" value={text} onChange={handleTextChange} />
+        </EditorField>
+      </>
       <AnnotationsHelp />
     </EditorRows>
   );
