+++ public/app/plugins/datasource/cloud-monitoring/components/GraphPeriod.tsx
@@ -1,11 +1,11 @@
 import React, { FunctionComponent } from 'react';
 
 import { SelectableValue } from '@grafana/data';
-import { Switch } from '@grafana/ui';
+import { EditorField, EditorRow, HorizontalGroup, Switch } from '@grafana/ui';
 
 import { GRAPH_PERIODS, SELECT_WIDTH } from '../constants';
 
-import { PeriodSelect, QueryEditorRow } from '.';
+import { PeriodSelect } from './index';
 
 export interface Props {
   refId: string;
@@ -16,8 +16,8 @@ export interface Props {
 
 export const GraphPeriod: FunctionComponent<Props> = ({ refId, onChange, graphPeriod, variableOptionGroup }) => {
   return (
-    <>
-      <QueryEditorRow
+    <EditorRow>
+      <EditorField
         label="Graph period"
         htmlFor={`${refId}-graph-period`}
         tooltip={
@@ -27,21 +27,23 @@ export const GraphPeriod: FunctionComponent<Props> = ({ refId, onChange, graphPe
           </>
         }
       >
-        <Switch
-          data-testid={`${refId}-switch-graph-period`}
-          value={graphPeriod !== 'disabled'}
-          onChange={(e) => onChange(e.currentTarget.checked ? '' : 'disabled')}
-        />
-        <PeriodSelect
-          inputId={`${refId}-graph-period`}
-          templateVariableOptions={variableOptionGroup.options}
-          current={graphPeriod}
-          onChange={onChange}
-          selectWidth={SELECT_WIDTH}
-          disabled={graphPeriod === 'disabled'}
-          aligmentPeriods={GRAPH_PERIODS}
-        />
-      </QueryEditorRow>
-    </>
+        <HorizontalGroup>
+          <Switch
+            data-testid={`${refId}-switch-graph-period`}
+            value={graphPeriod !== 'disabled'}
+            onChange={(e) => onChange(e.currentTarget.checked ? '' : 'disabled')}
+          />
+          <PeriodSelect
+            inputId={`${refId}-graph-period`}
+            templateVariableOptions={variableOptionGroup.options}
+            current={graphPeriod}
+            onChange={onChange}
+            selectWidth={SELECT_WIDTH}
+            disabled={graphPeriod === 'disabled'}
+            aligmentPeriods={GRAPH_PERIODS}
+          />
+        </HorizontalGroup>
+      </EditorField>
+    </EditorRow>
   );
 };
