+++ public/app/plugins/datasource/cloud-monitoring/components/GroupBy.tsx
@@ -1,13 +1,13 @@
 import React, { FunctionComponent, useMemo } from 'react';
 
 import { SelectableValue } from '@grafana/data';
-import { MultiSelect } from '@grafana/ui';
+import { EditorField, EditorFieldGroup, MultiSelect } from '@grafana/ui';
 
-import { INPUT_WIDTH, SYSTEM_LABELS } from '../constants';
+import { SYSTEM_LABELS } from '../constants';
 import { labelsToGroupedOptions } from '../functions';
 import { MetricDescriptor, MetricQuery } from '../types';
 
-import { Aggregation, QueryEditorRow } from '.';
+import { Aggregation } from './Aggregation';
 
 export interface Props {
   refId: string;
@@ -32,21 +32,22 @@ export const GroupBy: FunctionComponent<Props> = ({
   );
 
   return (
-    <QueryEditorRow
-      label="Group by"
-      tooltip="You can reduce the amount of data returned for a metric by combining different time series. To combine multiple time series, you can specify a grouping and a function. Grouping is done on the basis of labels. The grouping function is used to combine the time series in the group into a single time series."
-      htmlFor={`${refId}-group-by`}
-    >
-      <MultiSelect
-        inputId={`${refId}-group-by`}
-        width={INPUT_WIDTH}
-        placeholder="Choose label"
-        options={options}
-        value={query.groupBys ?? []}
-        onChange={(options) => {
-          onChange({ ...query, groupBys: options.map((o) => o.value!) });
-        }}
-      ></MultiSelect>
+    <EditorFieldGroup>
+      <EditorField
+        label="Group by"
+        tooltip="You can reduce the amount of data returned for a metric by combining different time series. To combine multiple time series, you can specify a grouping and a function. Grouping is done on the basis of labels. The grouping function is used to combine the time series in the group into a single time series."
+      >
+        <MultiSelect
+          inputId={`${refId}-group-by`}
+          width="auto"
+          placeholder="Choose label"
+          options={options}
+          value={query.groupBys ?? []}
+          onChange={(options) => {
+            onChange({ ...query, groupBys: options.map((o) => o.value!) });
+          }}
+        />
+      </EditorField>
       <Aggregation
         metricDescriptor={metricDescriptor}
         templateVariableOptions={variableOptionGroup.options}
@@ -54,7 +55,7 @@ export const GroupBy: FunctionComponent<Props> = ({
         groupBys={query.groupBys ?? []}
         onChange={(crossSeriesReducer) => onChange({ ...query, crossSeriesReducer })}
         refId={refId}
-      ></Aggregation>
-    </QueryEditorRow>
+      />
+    </EditorFieldGroup>
   );
 };
