+++ public/app/plugins/datasource/cloud-monitoring/components/LabelFilter.tsx
@@ -1,15 +1,9 @@
-import { flatten } from 'lodash';
-import React, { FunctionComponent, useCallback, useMemo } from 'react';
+import React, { FunctionComponent, useMemo } from 'react';
 
 import { SelectableValue, toOption } from '@grafana/data';
-import { Button, HorizontalGroup, Select, VerticalGroup } from '@grafana/ui';
-import { CustomControlProps } from '@grafana/ui/src/components/Select/types';
+import { AccessoryButton, EditorField, EditorList, EditorRow, HorizontalGroup, Select } from '@grafana/ui';
 
-import { SELECT_WIDTH } from '../constants';
 import { labelsToGroupedOptions, stringArrayToFilters } from '../functions';
-import { Filter } from '../types';
-
-import { QueryEditorRow } from '.';
 
 export interface Props {
   labels: { [key: string]: string[] };
@@ -18,137 +12,107 @@ export interface Props {
   variableOptionGroup: SelectableValue<string>;
 }
 
-const operators = ['=', '!=', '=~', '!=~'];
+interface Filter {
+  key: string;
+  operator: string;
+  value: string;
+  condition: string;
+}
 
-const FilterButton = React.forwardRef<HTMLButtonElement, CustomControlProps<string>>(
-  ({ value, isOpen, invalid, ...rest }, ref) => {
-    return <Button {...rest} ref={ref} variant="secondary" icon="plus" aria-label="Add filter"></Button>;
-  }
-);
-FilterButton.displayName = 'FilterButton';
+const DEFAULT_OPERATOR = '=';
+const DEFAULT_CONDITION = 'AND';
 
-const OperatorButton = React.forwardRef<HTMLButtonElement, CustomControlProps<string>>(({ value, ...rest }, ref) => {
-  return (
-    <Button {...rest} ref={ref} variant="secondary">
-      <span className="query-segment-operator">{value?.label}</span>
-    </Button>
-  );
-});
-OperatorButton.displayName = 'OperatorButton';
+const filtersToStringArray = (filters: Filter[]) =>
+  filters.flatMap(({ key, operator, value, condition }) => [key, operator, value, condition]).slice(0, -1);
+
+const operators = ['=', '!=', '=~', '!=~'].map(toOption);
 
 export const LabelFilter: FunctionComponent<Props> = ({
   labels = {},
   filters: filterArray,
-  onChange,
+  onChange: _onChange,
   variableOptionGroup,
 }) => {
-  const filters = useMemo(() => stringArrayToFilters(filterArray), [filterArray]);
+  const filters: Filter[] = useMemo(() => stringArrayToFilters(filterArray), [filterArray]);
   const options = useMemo(
     () => [variableOptionGroup, ...labelsToGroupedOptions(Object.keys(labels))],
     [labels, variableOptionGroup]
   );
 
-  const filtersToStringArray = useCallback((filters: Filter[]) => {
-    const strArr = flatten(filters.map(({ key, operator, value, condition }) => [key, operator, value, condition!]));
-    return strArr.slice(0, strArr.length - 1);
-  }, []);
+  const getOptions = ({ key = '', value = '' }: Partial<Filter>) => {
+    // Add the current key and value as options if they are manually entered
+    const keyPresent = options.some((op) => {
+      if (op.options) {
+        return options.some((opp) => opp.label === key);
+      }
+      return op.label === key;
+    });
+    if (!keyPresent) {
+      options.push({ label: key, value: key });
+    }
 
-  const AddFilter = () => {
+    const valueOptions = labels.hasOwnProperty(key)
+      ? [variableOptionGroup, ...labels[key].map(toOption)]
+      : [variableOptionGroup];
+    const valuePresent = valueOptions.some((op) => op.label === value);
+    if (!valuePresent) {
+      valueOptions.push({ label: value, value });
+    }
+
+    return { options, valueOptions };
+  };
+
+  const onChange = (items: Array<Partial<Filter>>) => {
+    const filters = items.map(({ key, operator, value, condition }) => ({
+      key: key || '',
+      operator: operator || DEFAULT_OPERATOR,
+      value: value || '',
+      condition: condition || DEFAULT_CONDITION,
+    }));
+    _onChange(filtersToStringArray(filters));
+  };
+
+  const renderItem = (item: Partial<Filter>, onChangeItem: (item: Filter) => void, onDeleteItem: () => void) => {
+    const { key = '', operator = DEFAULT_OPERATOR, value = '', condition = DEFAULT_CONDITION } = item;
+    const { options, valueOptions } = getOptions(item);
+
     return (
-      <Select
-        allowCustomValue
-        options={[variableOptionGroup, ...labelsToGroupedOptions(Object.keys(labels))]}
-        onChange={({ value: key = '' }) =>
-          onChange(filtersToStringArray([...filters, { key, operator: '=', condition: 'AND', value: '' }]))
-        }
-        menuPlacement="bottom"
-        renderControl={FilterButton}
-      />
+      <HorizontalGroup spacing="xs" width="auto">
+        <Select
+          aria-label="Filter label key"
+          formatCreateLabel={(v) => `Use label key: ${v}`}
+          allowCustomValue
+          value={key}
+          options={options}
+          onChange={({ value: key = '' }) => onChangeItem({ key, operator, value, condition })}
+        />
+        <Select
+          value={operator}
+          options={operators}
+          onChange={({ value: operator = DEFAULT_OPERATOR }) => onChangeItem({ key, operator, value, condition })}
+        />
+        <Select
+          aria-label="Filter label value"
+          placeholder="add filter value"
+          formatCreateLabel={(v) => `Use label value: ${v}`}
+          allowCustomValue
+          value={value}
+          options={valueOptions}
+          onChange={({ value = '' }) => onChangeItem({ key, operator, value, condition })}
+        />
+        <AccessoryButton aria-label="Remove" icon="times" variant="secondary" onClick={onDeleteItem} type="button" />
+      </HorizontalGroup>
     );
   };
 
   return (
-    <QueryEditorRow
-      label="Filter"
-      tooltip={
-        'To reduce the amount of data charted, apply a filter. A filter has three components: a label, a comparison, and a value. The comparison can be an equality, inequality, or regular expression.'
-      }
-      noFillEnd={filters.length > 1}
-    >
-      <VerticalGroup spacing="xs" width="auto">
-        {filters.map(({ key, operator, value, condition }, index) => {
-          // Add the current key and value as options if they are manually entered
-          const keyPresent = options.some((op) => {
-            if (op.options) {
-              return options.some((opp) => opp.label === key);
-            }
-            return op.label === key;
-          });
-          if (!keyPresent) {
-            options.push({ label: key, value: key });
-          }
-
-          const valueOptions = labels.hasOwnProperty(key)
-            ? [variableOptionGroup, ...labels[key].map(toOption)]
-            : [variableOptionGroup];
-          const valuePresent = valueOptions.some((op) => {
-            return op.label === value;
-          });
-          if (!valuePresent) {
-            valueOptions.push({ label: value, value });
-          }
-
-          return (
-            <HorizontalGroup key={index} spacing="xs" width="auto">
-              <Select
-                aria-label="Filter label key"
-                width={SELECT_WIDTH}
-                allowCustomValue
-                formatCreateLabel={(v) => `Use label key: ${v}`}
-                value={key}
-                options={options}
-                onChange={({ value: key = '' }) => {
-                  onChange(
-                    filtersToStringArray(
-                      filters.map((f, i) => (i === index ? { key, operator, condition, value: '' } : f))
-                    )
-                  );
-                }}
-              />
-              <Select
-                value={operator}
-                options={operators.map(toOption)}
-                onChange={({ value: operator = '=' }) =>
-                  onChange(filtersToStringArray(filters.map((f, i) => (i === index ? { ...f, operator } : f))))
-                }
-                menuPlacement="bottom"
-                renderControl={OperatorButton}
-              />
-              <Select
-                aria-label="Filter label value"
-                width={SELECT_WIDTH}
-                formatCreateLabel={(v) => `Use label value: ${v}`}
-                allowCustomValue
-                value={value}
-                placeholder="add filter value"
-                options={valueOptions}
-                onChange={({ value = '' }) =>
-                  onChange(filtersToStringArray(filters.map((f, i) => (i === index ? { ...f, value } : f))))
-                }
-              />
-              <Button
-                variant="secondary"
-                size="md"
-                icon="trash-alt"
-                aria-label="Remove"
-                onClick={() => onChange(filtersToStringArray(filters.filter((_, i) => i !== index)))}
-              ></Button>
-              {index + 1 === filters.length && Object.values(filters).every(({ value }) => value) && <AddFilter />}
-            </HorizontalGroup>
-          );
-        })}
-        {!filters.length && <AddFilter />}
-      </VerticalGroup>
-    </QueryEditorRow>
+    <EditorRow>
+      <EditorField
+        label="Filter"
+        tooltip="To reduce the amount of data charted, apply a filter. A filter has three components: a label, a comparison, and a value. The comparison can be an equality, inequality, or regular expression."
+      >
+        <EditorList items={filters} renderItem={renderItem} onChange={onChange} />
+      </EditorField>
+    </EditorRow>
   );
 };
