+++ public/app/plugins/datasource/cloud-monitoring/components/Preprocessor.tsx
@@ -1,12 +1,10 @@
 import React, { FunctionComponent, useMemo } from 'react';
 
 import { SelectableValue } from '@grafana/data';
-import { RadioButtonGroup } from '@grafana/ui';
+import { EditorField, RadioButtonGroup } from '@grafana/ui';
 
 import { getAlignmentPickerData } from '../functions';
 import { MetricDescriptor, MetricKind, MetricQuery, PreprocessorType, ValueTypes } from '../types';
-
-import { QueryEditorRow } from '.';
 
 const NONE_OPTION = { label: 'None', value: PreprocessorType.None };
 
@@ -19,7 +17,7 @@ export interface Props {
 export const Preprocessor: FunctionComponent<Props> = ({ query, metricDescriptor, onChange }) => {
   const options = useOptions(metricDescriptor);
   return (
-    <QueryEditorRow
+    <EditorField
       label="Pre-processing"
       tooltip="Preprocessing options are displayed when the selected metric has a metric kind of delta or cumulative. The specific options available are determined by the metic's value type. If you select 'Rate', data points are aligned and converted to a rate per time series. If you select 'Delta', data points are aligned by their delta (difference) per time series"
     >
@@ -31,8 +29,8 @@ export const Preprocessor: FunctionComponent<Props> = ({ query, metricDescriptor
         }}
         value={query.preprocessor ?? PreprocessorType.None}
         options={options}
-      ></RadioButtonGroup>
-    </QueryEditorRow>
+      />
+    </EditorField>
   );
 };
