+++ public/app/plugins/datasource/cloud-monitoring/components/Project.tsx
@@ -1,12 +1,9 @@
 import React, { useEffect, useMemo, useState } from 'react';
 
 import { SelectableValue } from '@grafana/data';
-import { Select } from '@grafana/ui';
+import { EditorField, Select } from '@grafana/ui';
 
-import { SELECT_WIDTH } from '../constants';
-import CloudMonitoringDatasource from '../datasource';
-
-import { QueryEditorRow } from '.';
+import CloudMonitoringDatasource from '../../datasource';
 
 export interface Props {
   refId: string;
@@ -35,9 +32,9 @@ export function Project({ refId, projectName, datasource, onChange, templateVari
   );
 
   return (
-    <QueryEditorRow label="Project" htmlFor={`${refId}-project`}>
+    <EditorField label="Project">
       <Select
-        width={SELECT_WIDTH}
+        width="auto"
         allowCustomValue
         formatCreateLabel={(v) => `Use project: ${v}`}
         onChange={({ value }) => onChange(value!)}
@@ -46,6 +43,6 @@ export function Project({ refId, projectName, datasource, onChange, templateVari
         placeholder="Select Project"
         inputId={`${refId}-project`}
       />
-    </QueryEditorRow>
+    </EditorField>
   );
 }
