+++ public/app/plugins/datasource/cloud-monitoring/components/QueryEditor.tsx
@@ -1,21 +1,16 @@
-import { css } from '@emotion/css';
 import React, { PureComponent } from 'react';
 
 import { QueryEditorProps, toOption } from '@grafana/data';
-import { config } from '@grafana/runtime';
-import { Button, EditorRows, Select } from '@grafana/ui';
+import { EditorRows } from '@grafana/ui';
 
-import { QUERY_TYPES, SELECT_WIDTH } from '../constants';
 import CloudMonitoringDatasource from '../datasource';
-import { CloudMonitoringQuery, EditorMode, MetricQuery, QueryType, SLOQuery, CloudMonitoringOptions } from '../types';
+import { CloudMonitoringQuery, MetricQuery, QueryType, SLOQuery, CloudMonitoringOptions } from '../types';
 
-import { MetricQueryEditor as ExperimentalMetricQueryEditor } from './Experimental/MetricQueryEditor';
-import { QueryHeader } from './Experimental/QueryHeader';
-import { SLOQueryEditor as ExperimentalSLOQueryEditor } from './Experimental/SLOQueryEditor';
 import { defaultQuery } from './MetricQueryEditor';
+import { QueryHeader } from './QueryHeader';
 import { defaultQuery as defaultSLOQuery } from './SLO/SLOQueryEditor';
 
-import { MetricQueryEditor, QueryEditorRow, SLOQueryEditor } from './';
+import { MetricQueryEditor, SLOQueryEditor } from './';
 
 export type Props = QueryEditorProps<CloudMonitoringDatasource, CloudMonitoringQuery, CloudMonitoringOptions>;
 
@@ -58,7 +53,7 @@ export class QueryEditor extends PureComponent<Props> {
       options: datasource.getVariables().map(toOption),
     };
 
-    return config.featureToggles.cloudMonitoringExperimentalUI ? (
+    return (
       <EditorRows>
         <QueryHeader
           query={query}
@@ -67,69 +62,6 @@ export class QueryEditor extends PureComponent<Props> {
           onChange={onChange}
           onRunQuery={onRunQuery}
         />
-        {queryType === QueryType.METRICS && (
-          <ExperimentalMetricQueryEditor
-            refId={query.refId}
-            variableOptionGroup={variableOptionGroup}
-            customMetaData={customMetaData}
-            onChange={(metricQuery: MetricQuery) => {
-              this.props.onChange({ ...this.props.query, metricQuery });
-            }}
-            onRunQuery={onRunQuery}
-            datasource={datasource}
-            query={metricQuery}
-          />
-        )}
-
-        {queryType === QueryType.SLO && (
-          <ExperimentalSLOQueryEditor
-            refId={query.refId}
-            variableOptionGroup={variableOptionGroup}
-            customMetaData={customMetaData}
-            onChange={(query: SLOQuery) => this.onQueryChange('sloQuery', query)}
-            onRunQuery={onRunQuery}
-            datasource={datasource}
-            query={sloQuery}
-          />
-        )}
-      </EditorRows>
-    ) : (
-      <EditorRows>
-        <QueryEditorRow
-          label="Query type"
-          fillComponent={
-            query.queryType !== QueryType.SLO && (
-              <Button
-                variant="secondary"
-                className={css`
-                  margin-left: auto;
-                `}
-                icon="edit"
-                onClick={() =>
-                  this.onQueryChange('metricQuery', {
-                    ...metricQuery,
-                    editorMode: metricQuery.editorMode === EditorMode.MQL ? EditorMode.Visual : EditorMode.MQL,
-                  })
-                }
-              >
-                {metricQuery.editorMode === EditorMode.MQL ? 'Switch to builder' : 'Edit MQL'}
-              </Button>
-            )
-          }
-          htmlFor={`${query.refId}-query-type`}
-        >
-          <Select
-            width={SELECT_WIDTH}
-            value={queryType}
-            options={QUERY_TYPES}
-            onChange={({ value }) => {
-              onChange({ ...query, sloQuery, queryType: value! });
-              onRunQuery();
-            }}
-            inputId={`${query.refId}-query-type`}
-          />
-        </QueryEditorRow>
-
         {queryType === QueryType.METRICS && (
           <MetricQueryEditor
             refId={query.refId}
