+++ public/app/plugins/datasource/cloud-monitoring/components/SLOQueryEditor.tsx
@@ -3,10 +3,10 @@ import React, { useMemo } from 'react';
 import { SelectableValue } from '@grafana/data';
 import { EditorField, EditorFieldGroup, EditorRow } from '@grafana/ui';
 
-import { ALIGNMENT_PERIODS, SLO_BURN_RATE_SELECTOR_NAME } from '../../constants';
-import CloudMonitoringDatasource from '../../datasource';
-import { alignmentPeriodLabel } from '../../functions';
-import { AlignmentTypes, CustomMetaData, SLOQuery } from '../../types';
+import { ALIGNMENT_PERIODS, SLO_BURN_RATE_SELECTOR_NAME } from '../constants';
+import CloudMonitoringDatasource from '../datasource';
+import { alignmentPeriodLabel } from '../functions';
+import { AlignmentTypes, CustomMetaData, SLOQuery } from '../types';
 
 import { AliasBy } from './AliasBy';
 import { LookbackPeriodSelect } from './LookbackPeriodSelect';
